Long counters (#2401)
* Used long for pitch accumulator and changed the bit shift setting to 31

* Used long for breath accumulator and changed the bit shift setting to 47

* Better comment for breath bit shift

* Breath is evaluated per fraction of a second and not per sample. I am a doofus.

+++ src/main/java/legend/core/audio/sequencer/Voice.java
@@ -30,7 +30,7 @@ final class Voice {
   /** playingNote.used_00 */
   private boolean used;
   /** voice.pitch */
-  private int sampleRate;
+  private long sampleRate;
   /** playingNote.noteNumber_02 */
   private int note;
   /** playingNote.velocityVolume_2c */
