+++ src/main/java/legend/core/audio/sequencer/assets/Breath.java
@@ -146,8 +146,8 @@ public final class Breath {
     return breath;
   }
 
-  public static int convert(final int value) {
-    return Math.round(BREATH_MAX_VALUE / (60 - 58 * value / 127.0f));
+  public static long convert(final int value) {
+    return Math.round(BREATH_MAX_VALUE / (60 - 58 * value / 127.0d));
   }
 
   public static Breath get(final FileData data, final int fileId, final int tableIndex) {
